Add handlers to background
